Drop support for build tags as this is non-scalable
Currently the min required version of libvirt is 1.2.2
and opt-in for certains APIs in newer versions can be
done with go build tags. This is non-scalable going
forward as there are a huge number of releases adding
APIs to libvirt. This will result in people building
the libvirt go bindings having to specify 10, 20, 30
or more build tags to 'go build'. Even worse, these
build tags are a global namespace, so if libvirt uses
a tag '1.2.14', you have to hope there's no other
library that also uses such a build tag. The latter
problem could be avoided by adding a 'libvirt' prefix
to the build tags, but this makes it even more verbose.

Instead, the bindings master branch will always track
the latest libvirt master branch. Apps needing build
compatibility with older libvirt versions will have to
use a previous tag/branch of the bindings.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>